+++ packages/runtime-core/src/remote/index.ts
@@ -5,7 +5,17 @@ import {
   ModuleInfo,
   GlobalModuleInfo,
 } from '@module-federation/sdk';
-import { Global, getInfoWithoutType, globalLoading } from '../global';
+import {
+  getShortErrorMsg,
+  RUNTIME_004,
+  runtimeDescMap,
+} from '@module-federation/error-codes';
+import {
+  Global,
+  getInfoWithoutType,
+  globalLoading,
+  CurrentGlobal,
+} from '../global';
 import {
   Options,
   UserOptions,
@@ -140,7 +150,7 @@ export class RemoteHandler {
     loadEntry: new AsyncHook<
       [
         {
-          createScriptHook: FederationHost['loaderHook']['lifecycle']['createScript'];
+          loaderHook: FederationHost['loaderHook'];
           remoteInfo: RemoteInfo;
           remoteEntryExports?: RemoteEntryExports;
         },
@@ -338,23 +348,12 @@ export class RemoteHandler {
       host.options.remotes,
       idRes,
     );
-
     assert(
       remoteSplitInfo,
-      `
-        Unable to locate ${idRes} in ${
-          host.options.name
-        }. Potential reasons for failure include:\n
-        1. ${idRes} was not included in the 'remotes' parameter of ${
-          host.options.name || 'the host'
-        }.\n
-        2. ${idRes} could not be found in the 'remotes' of ${
-          host.options.name
-        } with either 'name' or 'alias' attributes.
-        3. ${idRes} is not online, injected, or loaded.
-        4. ${idRes}  cannot be accessed on the expected.
-        5. The 'beforeRequest' hook was provided but did not return the correct 'remoteInfo' when attempting to load ${idRes}.
-      `,
+      getShortErrorMsg(RUNTIME_004, runtimeDescMap, {
+        hostName: host.options.name,
+        requestId: idRes,
+      }),
     );
 
     const { remote: rawRemote } = remoteSplitInfo;
@@ -468,14 +467,16 @@ export class RemoteHandler {
       const loadedModule = host.moduleCache.get(remote.name);
       if (loadedModule) {
         const remoteInfo = loadedModule.remoteInfo;
-        const key = remoteInfo.entryGlobalName as keyof typeof globalThis;
+        const key = remoteInfo.entryGlobalName as keyof typeof CurrentGlobal;
 
-        if (globalThis[key]) {
-          if (Object.getOwnPropertyDescriptor(globalThis, key)?.configurable) {
-            delete globalThis[key];
+        if (CurrentGlobal[key]) {
+          if (
+            Object.getOwnPropertyDescriptor(CurrentGlobal, key)?.configurable
+          ) {
+            delete CurrentGlobal[key];
           } else {
             // @ts-ignore
-            globalThis[key] = undefined;
+            CurrentGlobal[key] = undefined;
           }
         }
         const remoteEntryUniqueKey = getRemoteEntryUniqueKey(
@@ -493,7 +494,7 @@ export class RemoteHandler {
           ? composeKeyWithSeparator(remoteInfo.name, remoteInfo.buildVersion)
           : remoteInfo.name;
         const remoteInsIndex =
-          globalThis.__FEDERATION__.__INSTANCES__.findIndex((ins) => {
+          CurrentGlobal.__FEDERATION__.__INSTANCES__.findIndex((ins) => {
             if (remoteInfo.buildVersion) {
               return ins.options.id === remoteInsId;
             } else {
@@ -502,7 +503,7 @@ export class RemoteHandler {
           });
         if (remoteInsIndex !== -1) {
           const remoteIns =
-            globalThis.__FEDERATION__.__INSTANCES__[remoteInsIndex];
+            CurrentGlobal.__FEDERATION__.__INSTANCES__[remoteInsIndex];
           remoteInsId = remoteIns.options.id || remoteInsId;
           const globalShareScopeMap = getGlobalShareScope();
 
@@ -564,7 +565,7 @@ export class RemoteHandler {
               ];
             },
           );
-          globalThis.__FEDERATION__.__INSTANCES__.splice(remoteInsIndex, 1);
+          CurrentGlobal.__FEDERATION__.__INSTANCES__.splice(remoteInsIndex, 1);
         }
 
         const { hostGlobalSnapshot } = getGlobalRemoteInfo(remote, host);