+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @module-federation/modernjsapp
 
+## 0.1.64
+
+### Patch Changes
+
+- Updated dependencies [47fdbc2]
+  - @module-federation/enhanced@0.7.1
+
+## 0.1.63
+
+### Patch Changes
+
+- Updated dependencies [4eb09e7]
+  - @module-federation/enhanced@0.7.0
+
+## 0.1.62
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.16
+
+## 0.1.61
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.15
+
 ## 0.1.60
 
 ### Patch Changes