+++ apps/website-new/CHANGELOG.md
@@ -1,5 +1,25 @@
 # website-new
 
+## 1.0.6
+
+### Patch Changes
+
+- @module-federation/error-codes@0.7.1
+
+## 1.0.5
+
+### Patch Changes
+
+- e44d7fa: fix: parameter url for fetch options should not be necessary
+- Updated dependencies [849ef9c]
+  - @module-federation/error-codes@0.7.0
+
+## 1.0.4
+
+### Patch Changes
+
+- 283574b: feat(retry-plugin): allow fallback function to receive failed URL
+
 ## 1.0.3
 
 ### Patch Changes